requirements.txt
Flask==1.1.1
flask-restx==0.2.0
flask-sqlalchemy==2.4.4
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytest==6.1.1
pytest-flask==1.0.0
